Refactor util into a subpackage
