+++ setup.py
@@ -108,8 +108,8 @@ bridge_pyx = Extension('pynbody.bridge._bridge',
                      sources=['pynbody/bridge/_bridge.pyx'],
                      include_dirs=incdir)
 
-util_pyx = Extension('pynbody._util',
-                     sources=['pynbody/_util.pyx'],
+util_pyx = Extension('pynbody.util._util',
+                     sources=['pynbody/util/_util.pyx'],
                      include_dirs=incdir,
                      extra_compile_args=openmp_args,
                      extra_link_args=extra_link_args)
@@ -180,7 +180,7 @@ setup(name = 'pynbody',
       packages = ['pynbody', 'pynbody/analysis', 'pynbody/bc_modules', 'pynbody/array',
                   'pynbody/plot', 'pynbody/gravity', 'pynbody/chunk', 'pynbody/filt', 'pynbody/sph',
                   'pynbody/snapshot', 'pynbody/bridge', 'pynbody/halo', 'pynbody/halo/details',
-                  'pynbody/extern', 'pynbody/kdtree', 'pynbody/test_utils'],
+                  'pynbody/extern', 'pynbody/kdtree', 'pynbody/test_utils', 'pynbody/util'],
       package_data={'pynbody': ['default_config.ini'],
                     'pynbody/analysis': ['cmdlum.npz',
                                          'h1.hdf5',
